Start implementing client net code
We have a blocking issue. The current cell definition has texture information
baked into it. Obviously it's the clients responsibility for loading and
caring about textures, but the server needs to load the level and have
the cell state. We cannot just strip out the texture information since
currently we have collision detection code that checks for the presence
of a texture to determine whether there is a ceiling or floor in a given
cell.

TODO: Move to a cell definition that does not have texture information
but still gives the client enough information to determine what should
be rendered, and lets the server determine whether there is a ceiling or
a floor in any given cell.